+++ src/constants/base.js
@@ -4,6 +4,6 @@ export const MAIN_PANE_HEIGHT = '609px';
 export const columnWidths = {
   [searchResultListColumns.SELECT]: { min: 30, max: 30 },
   [searchResultListColumns.AUTH_REF_TYPE]: { min: 155, max: 166 },
-  [searchResultListColumns.HEADING_REF]: { min: 390, max: 440 },
-  [searchResultListColumns.HEADING_TYPE]: { min: 140, max: 127 },
+  [searchResultListColumns.HEADING_REF]: { min: 390, max: 435 },
+  [searchResultListColumns.HEADING_TYPE]: { min: 140, max: 145 },
 };