+++ src/views/BrowseView/BrowseView.js
@@ -1,37 +1,20 @@
 import {
-  useState,
   useContext,
   useMemo,
 } from 'react';
-import { useIntl } from 'react-intl';
 
-import { AppIcon } from '@folio/stripes/core';
 import {
-  Pane,
-  PaneMenu,
-} from '@folio/stripes/components';
-import { ExpandFilterPaneButton } from '@folio/stripes/smart-components';
-import {
-  AuthoritiesSearchPane,
-  SearchResultsList,
   AuthoritiesSearchContext,
-  SelectedAuthorityRecordContext,
   useAuthoritiesBrowse,
 } from '@folio/stripes-authority-components';
 
+import { AuthoritiesLookup } from '../../components';
 import {
-  MAIN_PANE_HEIGHT,
   PAGE_SIZE,
   PRECEDING_RECORDS_COUNT,
-  columnWidths,
 } from '../../constants';
 
 const BrowseView = () => {
-  const intl = useIntl();
-  const [isFilterPaneVisible, setIsFilterPaneVisible] = useState(true);
-
-  const toggleFilterPane = () => setIsFilterPaneVisible(!isFilterPaneVisible);
-
   const {
     filters,
     searchQuery,
@@ -41,7 +24,6 @@ const BrowseView = () => {
     searchInputValue,
     searchDropdownValue,
   } = useContext(AuthoritiesSearchContext);
-  const [, setSelectedAuthorityRecordContext] = useContext(SelectedAuthorityRecordContext);
 
   const {
     authorities,
@@ -60,41 +42,9 @@ const BrowseView = () => {
     precedingRecordsCount: PRECEDING_RECORDS_COUNT,
   });
 
-  const onSubmitSearch = e => {
-    if (e && e.preventDefault) {
-      e.preventDefault();
-      e.stopPropagation();
-    }
-
+  const onSubmitSearch = () => {
     setSearchQuery(searchInputValue);
     setSearchIndex(searchDropdownValue);
-    setSelectedAuthorityRecordContext(null);
-  };
-
-  const renderPaneSub = () => {
-    return (
-      <span>
-        {intl.formatMessage({
-          id: 'ui-plugin-find-authority.search-results-list.paneSub',
-        }, {
-          totalRecords,
-        })}
-      </span>
-    );
-  };
-
-  const renderResultsFirstMenu = () => {
-    if (isFilterPaneVisible) {
-      return null;
-    }
-
-    return (
-      <PaneMenu>
-        <ExpandFilterPaneButton
-          onClick={toggleFilterPane}
-        />
-      </PaneMenu>
-    );
   };
 
   const formattedAuthoritiesForView = useMemo(() => {
@@ -112,45 +62,20 @@ const BrowseView = () => {
   }, [authorities]);
 
   return (
-    <>
-      <AuthoritiesSearchPane
-        isFilterPaneVisible={isFilterPaneVisible}
-        toggleFilterPane={toggleFilterPane}
-        isLoading={isLoading}
-        onSubmitSearch={onSubmitSearch}
-        query={query}
-        height={MAIN_PANE_HEIGHT}
-      />
-      <Pane
-        id="authority-search-results-pane"
-        appIcon={<AppIcon app="marc-authorities" />}
-        defaultWidth="fill"
-        paneTitle={intl.formatMessage({ id: 'stripes-authority-components.meta.title' })}
-        paneSub={renderPaneSub()}
-        firstMenu={renderResultsFirstMenu()}
-        padContent={false}
-        noOverflow
-        height={MAIN_PANE_HEIGHT}
-      >
-        <SearchResultsList
-          authorities={formattedAuthoritiesForView}
-          columnWidths={columnWidths}
-          totalResults={totalRecords}
-          pageSize={PAGE_SIZE}
-          onNeedMoreData={handleLoadMore}
-          loading={isLoading}
-          loaded={isLoaded}
-          visibleColumns={['authRefType', 'headingRef', 'headingType']}
-          isFilterPaneVisible={isFilterPaneVisible}
-          toggleFilterPane={toggleFilterPane}
-          hasFilters={!!filters.length}
-          query={searchQuery}
-          hasNextPage={hasNextPage}
-          hasPrevPage={hasPrevPage}
-          hidePageIndices
-        />
-      </Pane>
-    </>
+    <AuthoritiesLookup
+      authorities={formattedAuthoritiesForView}
+      totalRecords={totalRecords}
+      searchQuery={searchQuery}
+      query={query}
+      isLoaded={isLoaded}
+      isLoading={isLoading}
+      hasFilters={!!filters.length}
+      hasNextPage={hasNextPage}
+      hasPrevPage={hasPrevPage}
+      hidePageIndices
+      onNeedMoreData={handleLoadMore}
+      onSubmitSearch={onSubmitSearch}
+    />
   );
 };