+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * Select a MARC authority record modal > Results list - Search. Refs UIPFAUTH-3
 * Select a MARC authority record modal > Results list - Browse. Refs UIPFAUTH-4
 * move stripes-authority-components to peerDependencies. Refs UIPFAUTH-9
+* Select a MARC authority record modal > View a MARC authority detail record. Refs UIPFAUTH-5